% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIres21rdf.R
\name{APIres21rdf}
\alias{APIres21rdf}
\title{API result to 1-row data frame}
\usage{
APIres21rdf(url)
}
\arguments{
\item{url}{}
}
\value{
A 1-row data frame containing the result of the API
}
\description{
API result to 1-row data frame
}
\examples{
out <- APIres21rdf("https://api.punkapi.com/v2/beers/1")
}
