% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_results}
\alias{check_results}
\title{Checks that the result is a 200 status code.}
\usage{
check_results(requete)
}
\arguments{
\item{requete}{requete sur l'API punk}
}
\value{
un booleen (status = 200 : TRUE / FALSE)
}
\description{
Checks that the result is a 200 status code.
}
\examples{
requete <- httr::GET("https://api.punkapi.com/v2/beers/1")
check_results(requete = requete)
}
