% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIres21rdf.R
\name{parse_result}
\alias{parse_result}
\title{API result to 1-row data frame}
\usage{
parse_result(url)
}
\arguments{
\item{url}{}
}
\value{
A 1-row data frame containing the result of the API
}
\description{
API result to 1-row data frame
}
\examples{
out <- parse_result("https://api.punkapi.com/v2/beers/1")
}
