% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beer.R
\name{print.beer}
\alias{print.beer}
\title{Printing method for beers}
\usage{
\method{print}{beer}(x, ...)
}
\arguments{
\item{x}{tibble got from a punk_xxx function}

\item{...}{}
}
\value{
print the beer characteristics
}
\description{
Printing method for beers
}
\examples{
res <- punk_by_id(id = 1)
print(res)
}
