% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punk_by_abv.R
\name{punk_by_abv}
\alias{punk_by_abv}
\title{punk_by_abv}
\usage{
punk_by_abv(abv_gt, abv_lt)
}
\arguments{
\item{abv_gt}{numeric in character}

\item{abv_lt}{numeric in character}
}
\value{
a tibble
}
\description{
punk_by_abv
}
\examples{
punk_by_abv(abv_gt = "3", abv_lt = "6")

}
