% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punk_by_id.R
\name{punk_by_id}
\alias{punk_by_id}
\alias{punk_by_ids}
\title{Get a Beer by its id}
\usage{
punk_by_id(id = NULL)

punk_by_ids(lst)
}
\arguments{
\item{id}{beer id}

\item{lst}{List of beer ids}
}
\value{
a dataframe with plenty of information about the chosen beer

A dataframe containing the API request data
}
\description{
Get a Beer by its id

Vectorized version of punk_by_id
}
\examples{
punk_by_id(id = 1)
punk_by_ids(lst = list(1,2,3,4))
}
