% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punk_randoms.R
\name{punk_random}
\alias{punk_random}
\alias{punk_randoms}
\title{Get a Random Beer}
\usage{
punk_random()

punk_randoms(n = 5)
}
\arguments{
\item{n}{number of element to return}
}
\value{
a dataframe with plenty of information about a random famous beer
}
\description{
Get a Random Beer
}
\examples{
punk_random()
punk_randoms(7)
}
