% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punk_randoms.R
\name{punk_randoms}
\alias{punk_randoms}
\title{punk_random}
\usage{
punk_randoms(path_api)
}
\arguments{
\item{path_api}{a valid API URL}
}
\value{
a dataframe with plenty of information about a random famous beer
}
\description{
punk_random
}
\examples{
punk_randoms(path_api = "https://api.punkapi.com/v2/beers/random")
}
