% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkapi-package.R
\docType{package}
\name{punkapi-package}
\alias{punkapi}
\alias{punkapi-package}
\title{punkapi: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})

}
\keyword{internal}
