% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punk_by_id.R, R/punk_randoms.R
\name{punk_by_id}
\alias{punk_by_id}
\alias{punk_random}
\alias{punk_randoms}
\title{Get a Beer by its id}
\usage{
punk_by_id(id = NULL)

punk_random()

punk_randoms(n = 5)
}
\arguments{
\item{id}{beer id}

\item{n}{number of element to return}
}
\value{
a dataframe with plenty of information about the chosen beer

a dataframe with plenty of information about a random famous beer
}
\description{
Get a Beer by its id

Get a Random Beer
}
\examples{
punk_by_id(id = 1)
punk_random()
punk_randoms(7)
}
